requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv
requests
Pillow==10.1.0
aiofiles==23.2.1
aiohttp==3.9.1
python-magic
pdf2image
# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
# Object Storage
boto3==1.34.0
# Background Workers
celery==5.3.4
redis==5.0.1
# Metrics & Observability
prometheus-client==0.19.0
# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1
# Streaming
sse-starlette==1.8.2
# Monetization
stripe==7.8.0
# Fine-tuning dependencies
transformers>=4.35.0
torch>=2.1.0
datasets>=2.14.0
peft>=0.6.0
accelerate>=0.24.0
faker>=19.0.0
tqdm>=4.66.0
python-dateutil>=2.8.0
# Receipt processing dependencies
opencv-python>=4.8.0
numpy>=1.24.0
pyyaml>=6.0
# Optional: for quantization
# optimum[onnxruntime]>=1.14.0
# Optional: for TEDS evaluation
# teds>=1.0.0
# CAMEL-AI integration
camel-ai[all]>=0.2.5
fastmcp
# Unsloth integration (install from GitHub: pip install "unsloth @ git+https://github.com/unslothai/unsloth.git")
# trl>=0.7.0  # For SFTTrainer (uncomment if using Unsloth training)
# bitsandbytes>=0.41.0  # For QLoRA quantization (uncomment if using QLoRA)
